Improve `Order` creation

+++ src/api/gdax/rest.rs
@@ -46,7 +46,7 @@ impl Client {
             endpoint,
         );
 
-        let timestamp = timestamp_ms() / 1000;
+        let timestamp = (timestamp_ms() as f64) / 1000.;
 
         let mut signer = Signer::new(MessageDigest::sha256(), &keys.secret_key).unwrap();
         let what = format!("{}{}/{}{}", timestamp, method, endpoint, body);
